+++ docs/release-notes.rst
@@ -20,6 +20,11 @@
 RIC Dashboard Release Notes
 ===========================
 
+Version 1.0.4, 29 May 2019
+--------------------------
+* Add ANR xApp neighbor cell relation table
+* Drop the pendulum xApp control screen
+
 Version 1.0.3, 28 May 2019
 --------------------------
 * Add AC xApp controller to backend